src/main/java/cd/go/authentication/passwordfile/executor/GetCapabilitiesExecutor.java
package cd.go.authentication.passwordfile.executor;

import cd.go.authentication.passwordfile.model.Capabilities;
import cd.go.authentication.passwordfile.model.SupportedAuthType;
import com.thoughtworks.go.plugin.api.response.DefaultGoPluginApiResponse;
import com.thoughtworks.go.plugin.api.response.GoPluginApiResponse;

import static com.thoughtworks.go.plugin.api.response.DefaultGoApiResponse.SUCCESS_RESPONSE_CODE;

public class GetCapabilitiesExecutor {

    public GoPluginApiResponse execute() {
        Capabilities capabilities = getCapabilities();
        return new DefaultGoPluginApiResponse(SUCCESS_RESPONSE_CODE, capabilities.toJSON());
    }

    Capabilities getCapabilities() {
        return new Capabilities(SupportedAuthType.Password, true, false, false);
    }
}
